Fix bug in verifyDoc when doc is not found.
When the document being verified has not been added the chaincode
returns an empty result and the JSON.parse was crashing on it.
Now the verifyDoc checks that the length of the returned result is
not 0 before processing it.  Otherwise it returns an error of
'Document not found'.

Also fix up the alert message popups so that they will display
the returned results.  A better display of the returned document
receipt will be needed in the future.