Check to see if ./tmp exists and create it
If the ./tmp directory that is being used to store the keyValStore
on the local machine doesn't exist the SDK will fail.  This code
checks is if it is there and creates it if not.  Then a final check
is done to see if it is there and if not the application will fail.