Make the /member REST endpoint more robust
Add checks for if the user/member exists and if it is check
to see that it is registered and enrolled.  If any of the checks
fail an error message will be send back to the browser.  Otherwise
an OK message along with the user name is sent back to the browser.